Another Update to the Village
I updated the village further, added a workshop, field kitchen, and fighting arena to sort out disputes in RP and for passing time. Also put the commissary in the former chief's room and added a soviet outfit there to indicate where the tribe got the spaceman things present in the village.