Made the gpstag change use a signal instead, cleaned up a few more artifacts from earlier versions of this code.
Adds a new file called gps.dm, for the sole reason of adding the signal registration of when a signal source is renamed. This should allow in the future for some easy handling of name changes if we want something that's not a GPS item to get its gpstag changed anyway.